漏洞修复 【CVE-2023-45199】

+++ library/ssl_tls12_client.c
@@ -1714,7 +1714,7 @@ static int ssl_parse_server_ecdh_params(mbedtls_ssl_context *ssl,
                                         unsigned char *end)
 {
     uint16_t tls_id;
-    uint8_t ecpoint_len;
+    size_t ecpoint_len;
     mbedtls_ssl_handshake_params *handshake = ssl->handshake;
     psa_ecc_family_t ec_psa_family = 0;
     size_t ec_bits = 0;
